Add argument to Mockito ThenThrow calls
I'd been using it with no argument since it seemed to work, but after
having some trouble getting a test to work right, I figured out that
actually causes something to be thrown whenever there's a catch block to
catch it, regardless of whether the stubbed method is actually called in
the try block. Very unexpected behavior and I haven't been able to find
any documentation regarding this.